+++ android/src/main/java/com/brentvatne/common/VideoTrack.java
@@ -1,0 +1,12 @@
+package com.brentvatne.common;
+
+public class VideoTrack
+{
+    public int m_width = 0;
+    public int m_height = 0;
+    public int m_bitrate = 0;
+    public String m_codecs = "";
+    public int m_id = -1;
+    public String m_trackId = "";
+    public boolean m_isSelected = false;
+}