+++ android/src/main/java/com/brentvatne/exoplayer/ExoPlayerView.java
@@ -4,7 +4,6 @@ import android.annotation.TargetApi;
 import android.content.Context;
 import androidx.core.content.ContextCompat;
 import android.util.AttributeSet;
-import android.util.Log;
 import android.util.TypedValue;
 import android.view.Gravity;
 import android.view.SurfaceView;
@@ -29,7 +28,6 @@ import com.google.android.exoplayer2.video.VideoSize;
 
 import java.util.List;
 
-@TargetApi(16)
 public final class ExoPlayerView extends FrameLayout implements AdViewProvider {
 
     private View surfaceView;